+++ src/entries/background/index.js
@@ -1,5 +1,5 @@
 import { SETTINGS_KEY, getSettings } from "../../lib/db.js";
-import { MENU_ID, getDefaultSettings } from "../../config/background.js";
+import { getDefaultSettings } from "../../config/background.js";
 import { initVisitTracking } from "./visits.js";
 
 const state = {
@@ -30,33 +30,6 @@ chrome.storage.onChanged.addListener((changes, area) => {
 });
 
 loadSettings();
-
-function openLibrary() {
-  if (chrome.runtime?.openOptionsPage) {
-    chrome.runtime.openOptionsPage();
-    return;
-  }
-  const url = chrome.runtime.getURL("src/entries/options/index.html");
-  chrome.tabs.create({ url });
-}
-
-chrome.runtime.onInstalled.addListener(() => {
-  try {
-    chrome.contextMenus.create({
-      id: MENU_ID,
-      title: "Open CamKeeper Library",
-      contexts: ["action", "browser_action"],
-    });
-  } catch (error) {
-    // Ignore context menu creation errors in unsupported contexts.
-  }
-});
-
-chrome.contextMenus.onClicked.addListener((info) => {
-  if (info.menuItemId !== MENU_ID) return;
-  openLibrary();
-});
-
 
 chrome.tabs.onActivated.addListener((info) => visits.onTabActivated(info));
 chrome.tabs.onUpdated.addListener((tabId, changeInfo) => visits.onTabUpdated(tabId, changeInfo));